requirements.txt
aiofiles==24.1.0
azure-ai-inference==1.0.0b6
azure-ai-projects==1.0.0b4
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
idna==3.10
isodate==0.7.2
msal==1.31.1
msal-extensions==1.2.0
portalocker==2.10.1
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
six==1.17.0
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.3.0
